backend/requirements.txt
alembic==1.15.2
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cryptography==45.0.2
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mysqlclient==2.2.7
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.0
requests==2.32.4
SQLAlchemy==2.0.41
typing_extensions==4.13.2
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
